Move message classes to subdirectory; remove 'message_' from filenames.
