+++ lib/dnsruby/message/encoder.rb
