+++ lib/dnsruby/message/header.rb
